+++ src/main/java/nextstep/theme/controller/ThemeController.java
@@ -1,21 +1,13 @@
 package nextstep.theme.controller;
 
 import lombok.RequiredArgsConstructor;
-import nextstep.auth.support.LoginRequired;
 import nextstep.theme.model.Theme;
-import nextstep.theme.model.ThemeRequest;
 import nextstep.theme.service.ThemeService;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import javax.validation.Valid;
-import java.net.URI;
 import java.util.List;
 
 @RestController
@@ -24,24 +16,9 @@ import java.util.List;
 public class ThemeController {
     private final ThemeService themeService;
 
-    @LoginRequired
-    @PostMapping
-    public ResponseEntity<Void> createTheme(@Valid @RequestBody ThemeRequest themeRequest) {
-        Long id = themeService.create(themeRequest);
-        return ResponseEntity.created(URI.create("/themes/" + id)).build();
-    }
-
     @GetMapping
     public ResponseEntity<List<Theme>> showThemes() {
         List<Theme> results = themeService.findAll();
         return ResponseEntity.ok().body(results);
-    }
-
-    @LoginRequired
-    @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteTheme(@PathVariable Long id) {
-        themeService.delete(id);
-
-        return ResponseEntity.noContent().build();
     }
 }
